make package work again
