+++ packages/web-service-worker/package.json
@@ -10,24 +10,99 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "17.1.0",
-    "@rollup/plugin-typescript": "^9.0.2",
-    "rollup": "2.78.0",
-    "rollup-plugin-babel": "^4.4.0",
-    "rollup-plugin-modify": "^3.0.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-polyfill-node": "^0.11.0",
-    "ts-loader": "6.2.1",
-    "workbox-cli": "^6.5.4",
-    "workbox-window": "^6.5.4"
+    "@babel/core": "7.20.2",
+    "@babel/polyfill": "7.12.1",
+    "@babel/preset-env": "7.20.2",
+    "@babel/register": "7.18.9",
+    "@cucumber/cucumber": "7.3.2",
+    "@cucumber/messages": "17.1.1",
+    "@cucumber/pretty-formatter": "1.0.0",
+    "@noble/hashes": "1.1.3",
+    "@playwright/test": "1.28.0",
+    "@rollup-extras/plugin-copy": "1.5.3",
+    "@rollup/plugin-alias": "4.0.2",
+    "@rollup/plugin-babel": "6.0.2",
+    "@rollup/plugin-commonjs": "23.0.2",
+    "@rollup/plugin-html": "1.0.1",
+    "@rollup/plugin-inject": "5.0.2",
+    "@rollup/plugin-json": "5.0.1",
+    "@rollup/plugin-node-resolve": "15.0.1",
+    "@rollup/plugin-terser": "0.1.0",
+    "@types/jest": "27.5.2",
+    "@types/jest-axe": "3.5.5",
+    "@types/lodash-es": "4.17.6",
+    "@types/node": "16.18.3",
+    "@types/node-fetch": "2.6.2",
+    "autoprefixer": "10.4.13",
+    "babel-core": "7.0.0-bridge.0",
+    "babel-jest": "29.3.1",
+    "commander": "8.3.0",
+    "core-js": "3.26.1",
+    "cucumber-html-reporter": "5.5.0",
+    "depcheck": "1.4.3",
+    "ejs": "3.1.8",
+    "eslint": "8.27.0",
+    "flush-promises": "1.0.2",
+    "focus-trap": "6.9.4",
+    "git-repo-info": "2.1.1",
+    "jest": "29.3.1",
+    "jest-axe": "5.0.1",
+    "jest-environment-jsdom": "29.3.1",
+    "jest-fetch-mock": "3.0.3",
+    "jest-mock-axios": "4.6.2",
+    "jest-mock-extended": "3.0.1",
+    "join-path": "1.1.1",
+    "license-checker-rseidelsohn": "3.1.0",
+    "lodash": "4.17.21",
+    "node-fetch": "2.6.7",
+    "pino": "7.11.0",
+    "pino-pretty": "7.6.1",
+    "playwright": "1.28.0",
+    "postcss": "8.4.19",
+    "qs": "6.11.0",
+    "react": "17.0.2",
+    "requirejs": "2.3.6",
+    "rollup": "3.3.0",
+    "rollup-plugin-delete": "2.0.0",
+    "rollup-plugin-gzip": "3.1.0",
+    "rollup-plugin-livereload": "2.0.5",
+    "rollup-plugin-modify": "3.0.0",
+    "rollup-plugin-polyfill-node": "0.11.0",
+    "rollup-plugin-postcss": "4.0.2",
+    "rollup-plugin-progress": "1.1.2",
+    "rollup-plugin-serve": "2.0.1",
+    "rollup-plugin-ts": "3.0.2",
+    "rollup-plugin-visualizer": "5.8.3",
+    "ts-jest": "29.0.3",
+    "ts-node": "10.9.1",
+    "tslib": "2.4.1",
+    "typescript": "4.9.3",
+    "url-search-params-polyfill": "8.1.1",
+    "wait-for-expect": "3.0.2"
   },
   "dependencies": {
+    "@babel/polyfill": "^7.12.1",
+    "@noble/hashes": "^1.1.3",
     "@rollup/plugin-alias": "3.1.9",
+    "axios": "^1.1.3",
+    "browserify": "^17.0.0",
+    "noble": "^1.9.1",
     "owncloud-sdk": "~3.0.0-alpha.17",
     "path": "^0.12.7",
+    "pbkdf2": "link:@noble/hashes/pbkdf2",
+    "promise": "^8.3.0",
     "qs": "^6.10.3",
+    "sax": "^1.2.4",
+    "sha512": "link:@noble/hashes/sha512",
     "typescript": "^4.8.4",
-    "workbox-precaching": "^6.5.4"
+    "utf8": "^3.0.0",
+    "uuid": "^9.0.0",
+    "workbox": "^0.0.0",
+    "workbox-cli": "^6.5.4",
+    "workbox-precaching": "^6.5.4",
+    "xhr-shim": "^0.1.3",
+    "xml-js": "^1.6.11",
+    "xmlhttprequest": "^1.8.0"
   },
   "keywords": [],
   "author": "",
